debug

+++ pr_page_fault_pinned_second_iteration_random_K40/pr_page_fault_pinned_second_iteration_random_K40.cu
@@ -14,14 +14,14 @@
 
 //typedef unsigned char byte;
 
-void shuffle(int *array, size_t n)
+void shuffle(int *array, long long int n)
 {
     if (n > 1) 
     {
-        size_t i;
+        long long int i;
         for (i = 0; i < n - 1; i++) 
         {
-          size_t j = i + rand() / (RAND_MAX / (n - i) + 1);
+          long long int j = i + rand() / (RAND_MAX / (n - i) + 1);
           int t = array[j];
           array[j] = array[i];
           array[i] = t;
